feat: add best time in table of winners
