+++ src/modules/markup/base/base.css
@@ -19,14 +19,17 @@ main {
     flex-direction: column;
     gap: 1rem;
 }
+
 button {
     border: 0px;
     border-radius: 5rem;
     margin: .2rem 0 0 .5rem;
 }
+
 .wrapper-cars__page {
     display: none;
 }
+
 .wrapper-cars {
     display: flex;
     flex-direction: column;
@@ -63,7 +66,7 @@ button {
 }
 
 .title-winner {
-position: fixed;
+    position: fixed;
     left: 0;
     right: 0;
     /* bottom: 0; */
@@ -73,4 +76,54 @@ position: fixed;
     font-size: 3rem;
     box-shadow: 0rem 0rem 4rem 1px #d53c3e;
     color: #b84b25;
+}
+
+
+
+
+/* -------------------------------Carage ---------------------------------
+ */
+.container-garage {
+    display: none;
+}
+
+.container-garage__wrapper-list-winners {
+    font-size: 1.2rem;
+
+}
+
+.wrapper-list-winners__wrapper-winners {
+    padding: 0;
+    margin: 0;
+    flex-basis: 100%;
+
+
+}
+
+.wrapper-list-winners__names-columns {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr) 2fr;
+    background-color: #10101094;
+    font-weight: 800;
+    border-radius: 1rem;
+    padding: 0 0 0 0.3rem;
+
+}
+
+.wrapper-winners__winner {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr) 2fr;
+    align-items: center;
+    padding: 0 0 0 0.3rem;
+}
+
+.winner__car-img {
+    height: 2rem;
+    width: 3rem;
+    flex-basis: 100%;
+    /* background-image: url(../../../assets/svg/car.svg); */
+    mask: url(../../../assets/svg/car.svg);
+    mask-repeat: no-repeat;
+    background-color: greenyellow;
+
 }
